Big refactoring.
