+++ YasEngineSoftware/InputOutputHandler.hpp
@@ -16,15 +16,11 @@ namespace YasInOut
 	struct MousePositionChangeInformation {
 		double x = 0.0;
 		double y = 0.0;
-		//double oldX = 0.0;
-		//double oldY = 0.0;
 		double angle = 0.0;
 		bool mouseMoved = false;
 		bool leftMouseButton;
 		bool rightMouseButton;
 	};
 }
-
-
 
 #endif
